Remove abstractions of MMapDirectory and MR-JAR support and move it to main sourceSet
