+++ gradle/generation/extract-jdk-apis.gradle
@@ -20,7 +20,7 @@ def resources = scriptResources(buildscript)
 configure(project(":lucene:core")) {
   ext {
     apijars = layout.projectDirectory.dir("src/generated/jdk")
-    mrjarJavaVersions = [ 21 ]
+    mrjarJavaVersions = [ 23 ]
   }
   
   configurations {
