+++ lucene/core/src/java/org/apache/lucene/store/MemorySegmentIndexInput.java
