+++ lucene/core/src/java/org/apache/lucene/store/NativeAccess.java
