+++ lucene/core/src/java/org/apache/lucene/store/RefCountedSharedArena.java
