+++ lucene/core/src/java23/org/apache/lucene/internal/vectorization/Lucene99MemorySegmentFlatVectorsScorer.java
