+++ lucene/core/src/java23/org/apache/lucene/internal/vectorization/PanamaVectorConstants.java
