+++ lucene/core/src/test/org/apache/lucene/store/TestMMapDirectory.java
@@ -25,7 +25,6 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Random;
 import java.util.concurrent.Callable;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Executors;
@@ -227,10 +226,7 @@ public class TestMMapDirectory extends BaseDirectoryTestCase {
         }
       }
 
-      if (!(dir.attachment instanceof ConcurrentHashMap<?, ?> map)) {
-        throw new AssertionError("unexpected attachment: " + dir.attachment);
-      }
-      assertEquals(0, map.size());
+      assertEquals(0, dir.arenas.size());
     }
   }
 
@@ -279,10 +275,7 @@ public class TestMMapDirectory extends BaseDirectoryTestCase {
         closeable.close();
       }
 
-      if (!(dir.attachment instanceof ConcurrentHashMap<?, ?> map)) {
-        throw new AssertionError("unexpected attachment: " + dir.attachment);
-      }
-      assertEquals(0, map.size());
+      assertEquals(0, dir.arenas.size());
     }
   }
