--- lucene/licenses/asm-9.6.jar.sha1
@@ -1,1 +1,0 @@
-aa205cf0a06dbd8e04ece91c0b37c3f5d567546a
