--- lucene/licenses/asm-commons-9.6.jar.sha1
@@ -1,1 +1,0 @@
-f1a9e5508eff490744144565c47326c8648be309
