--- lucene/licenses/asm-tree-9.6.jar.sha1
@@ -1,1 +1,0 @@
-c0cdda9d211e965d2a4448aa3fd86110f2f8c2de
